% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRecode.R
\name{data.recode}
\alias{data.recode}
\title{data.recode}
\usage{
data.recode(x, continuous = NULL, ordinal = NULL, nominal = NULL,
  count = NULL, order = NULL, concomitant = NULL)
}
\arguments{
\item{x}{input data}

\item{continuous}{a vecter indicates the columns of continuous variables.}

\item{ordinal}{a vecter indicates the columns of ordinal variables.}

\item{nominal}{a vecter indicates the columns of nominal variables.}

\item{count}{a vecter indicates the columns of count variables.}

\item{order}{a list indicates the order of each ordinal variable.}
}
\value{
a list with components
     \item{data}{numeric data matrix with the order of continuous, ordinal,
     nominal, count variables}
     \item{ppdim}{vector of nominal variable-wise numbers of categories.}
     \item{con.name/ord.name/count.name}{name of continuous/ordinal/count variables}
     \item{ord.levels/nom.levels}{lists of levels for each ordinal/nominal variables}
     \item{con.len/ord.len/nom.len/count.len}{number of each kinds of variables}
     \item{ordata}{original data after exchange columns}
}
\description{
Prepare data to be used in function \code{\link{mcmixed}}. This function
recodes discrete/categorical variables to standard number.
}

