% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmclust.R
\name{fmclust}
\alias{fmclust}
\title{fmclust}
\usage{
fmclust(x, continuous = NULL, ordinal = NULL, nominal = NULL,
  count = NULL, order = NULL, nClust = 2:5, concomitant = NULL,
  diagonal = TRUE)
}
\arguments{
\item{x}{input dataframe}

\item{continuous}{a vecter indicates the columns of continuous variables.}

\item{ordinal}{a vecter indicates the columns of ordinal variables.}

\item{nominal}{a vecter indicates the columns of nominal variables.}

\item{count}{a vecter indicates the columns of count variables.}

\item{order}{a list indicates the order of each ordinal variable.}

\item{nClust}{a vecter of number of clusters}

\item{diagonal}{if assume local independence between continuous variable. default TRUE}
}
\value{
an object of class \code{stepFlexmix} containing the best models with respect to the
log likelihood for the different number of components in a slot if \code{length(k)>1},
else directly an object of class \code{flexmix}.
}
\description{
\code{fmclust} fits a finite mixture clustering model where some variables are continuous and
modelled within the mixture components by Gaussian distributions, some variables are nominal
and modelled by independent multinomial distributions, and some variables are count and
modelled by independent poisson distributions. The fit is by maximum likelihood estimation
computed with the EM algorithm, based on function \code{initFlexmix} from package \code{flexmix}
}
\examples{
# simulate data
gg <- simuData(3, size = c(100,100,100), cont = 3, ord = 2,  nom =3, count=3,
     center = center, var = var, centerOrd = centerOrd, varOrd = varOrd, podim = podim,
     paraNom = paraNom, ppdim = ppdim, lambda = lambda)

# run models
mds <- fmclust(gg$data[,-1], continuous = 1:3, ordinal = 4:5, nominal = 6:8, count = 9:11)

# get the best model by BIC
m <- getModel(mds, which="BIC")
m

# compare with the real cluster
t <- table(gg$data[,1],m@cluster)
t

}

