% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmixed.R
\name{mcmixed}
\alias{mcmixed}
\title{mcmixed}
\usage{
mcmixed(formula = . ~ ., continuous = 0, ordinal = 0, nominal = 0,
  count.poi = 0, ppdim, diagonal = TRUE, pred.ordinal = FALSE,
  printlik = FALSE)
}
\arguments{
\item{formula}{use default when do clustering}

\item{continuous}{number of continuous variables}

\item{ordinal}{number of ordinal variables}

\item{nominal}{number of nominal variables}

\item{count.poi}{number of count variables}

\item{diagonal}{whether assume local independence. default TRUE (FALSE set full variance in continuous data)}

\item{printlik}{whether print the loglikelihood}
}
\description{
This is a M-step used to do clustering in flexmix, it could use to cluster mixed-mode (continuous/ordinal/nominal/count) data
}

