% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{profile}
\usage{
profile(mod, data)
}
\arguments{
\item{mod}{object of class \code{flexmix}}

\item{data}{data obeject of \code{data.recode}}
}
\value{
a list with component
   \item{dfPlt}{data frame used in profile plot (without split of continuous&ordinal)}
   \item{dfPro}{data frame of profile}
   \item{dfPrM}{data frame of probMean}
}
\description{
profile
}

