% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simuData}
\alias{simuData}
\title{simuData}
\usage{
simuData(ngroup, size, cont = 0, ord = 0, nom = 0, count = 0,
  center = NULL, var = NULL, centerOrd = NULL, varOrd = NULL,
  podim = NULL, paraNom = NULL, ppdim = NULL, lambda = NULL,
  seed = 123)
}
\arguments{
\item{ngroup}{integer indicates the number of groups}

\item{size}{vector of sizes of each groups}

\item{center}{a matrix of means for each continuous variable in each group (ngroup * cont)}

\item{var}{a matrix of variances for each continuous variable in each group (ngroup * cont)}

\item{centerOrd}{a matrix of means for each ordinal variable in each group (ngroup * ord)}

\item{varOrd}{a matrix of variances for each ordinal variable in each group (ngroup * ord)}

\item{podim}{vector of number of categories for each ordinal data}

\item{cont/ord/nom/count}{number of continuous/ordinal/nominal/count(poisson) variables}

\item{paraNom:}{a list of matrix. In each matrix, element (i,j) indicates the probabilities of}
}
\value{
a list with component
\item{data}{simulated data}
\item{statistics}{parameters used in simulation}
}
\description{
Simulate data and related parameters if not defined.
}

